Update filterMovies with conditional
