+++ src/App/App.js
@@ -4,7 +4,6 @@ import { Routes, Route } from "react-router-dom";
 import "../App/App.css";
 // import movieData from "../Data/movieData";
 import { getAllMovies } from "../apiCalls";
-import SearchMovies from "../Components/Search/search"
 
 import AllMovies from "../Components/AllMovies/AllMovies";
 import MovieBlowUp from "../Components/MovieBlowUp/MovieBlowUp";
@@ -12,8 +11,6 @@ import MovieBlowUp from "../Components/MovieBlowUp/MovieBlowUp";
 
 function App() {
   const [movies, setMovies] = useState([]);
-  const [filteredMovies, setFilteredMovies] = useState([]) 
-  // move to AllMovies
   const [error, setError] = useState("");
 
   useEffect(() => {
@@ -35,24 +32,6 @@ function App() {
       });
   };
 
-// move this into All movies
-const filterMovies = (searchTerms) => {   
-  const  filteredMovies = movies.filter((movie) => {
-    const searchTermString = searchTerms.toLowerCase().split(" ").join("")
-    return movie.title.toLowerCase().split(" ").join("").includes(searchTermString)
-  })
-  if (filteredMovies.length === 0 ) {
-    setFilteredMovies( movies)
-  } else {
-    setFilteredMovies(filteredMovies)
-  }
-  // if filteredMovies.length === 0, we want to reassign filteredMovies to the spread of movies
-
-  // set movie state again here?
-  // setup the route for search and pass state of filtered movie as prop
-  // have a way to change the url > when you start typing 
-}
-
   const clearError = () => {
     setError("");
   };
@@ -61,13 +40,10 @@ const filterMovies = (searchTerms) => {
     <main className="App">
       <nav>
         <h1 className="main-header">Rancid Tomatillos</h1>
-        {/* Move form to all movies  */}
-        <SearchMovies filterMovies={filterMovies}/>
       </nav>
       {error && <h2 className="movies-error">{error}</h2>}
       <Routes>
         <Route path="/" element={<AllMovies moviesData={movies} />} />
-        {/* <Route path="/search" element={<AllMovies moviesData={filteredMovies} />} /> */}
         <Route
           path="/movie/:id"
           element={
