+++ src/Components/AllMovies/AllMovies.js
@@ -1,6 +1,9 @@
 import "./AllMovies.css";
 import MovieCard from "../MovieCards/MovieCards";
 import PropTypes from "prop-types";
+import SearchMovies from "../Search/search"
+import { useState, useEffect } from "react"
+
 
 // add filtering/search functionality here in APP component
 // set state for search
@@ -11,15 +14,34 @@ import PropTypes from "prop-types";
     // - use find() to search through the array to find the first word from 
     // the title that starts with a user's search value. 
     // should not be case sensitive: toLowerCase()?
-// - startsWith()?
 // - set the state to the result of this filtering
 
 
 function AllMovies({ moviesData }) {
+  const [filteredMovies, setFilteredMovies] = useState([]) 
+  
+  useEffect(() => {
+    setFilteredMovies([...moviesData])
+}, [moviesData])
 
+  const filterMovies = (searchTerms) => {   
+    console.log("search terms", searchTerms)
+    if(searchTerms !== undefined) {
+      const searchResults = moviesData.filter((movie) => {
+        const searchTermString = searchTerms.toLowerCase().split(" ").join("")
+        return movie.title.toLowerCase().split(" ").join("").includes(searchTermString)
+      })
+      console.log("SEARCH RESULTS", searchResults)
+    if (searchResults.length === 0 ) {
+      setFilteredMovies([...moviesData])
 
+    } else {
+      setFilteredMovies(searchResults)
+    }
+    }
+  }
 
-  const movieCards = moviesData.map((movie) => {
+  const movieCards = filteredMovies.map((movie) => {
     return (
       <MovieCard
         imagePath={movie.poster_path}
@@ -27,11 +49,15 @@ function AllMovies({ moviesData }) {
         title={movie.title}
         id={movie.id}
         key={movie.id}
-        // handleSingleMovie={handleSingleMovie}
       />
     );
   });
-  return <div className="movie-container">{movieCards}</div>;
+  return (
+  <div>
+    <SearchMovies filterMovies={filterMovies}/>
+        <div className="movie-container">{movieCards}</div>
+  </div>
+    )
 }
 
 AllMovies.propTypes = {
